+++ nix/haskell-parsers/parser_tests.nix
@@ -17,6 +17,15 @@ let
         '';
         expected = [ "foo" "bar" ];
       };
+
+      # Handles cases where cabal.project does not end with newline
+      testEOF = {
+        expr = eval ''
+          packages:
+            foo
+            bar'';
+        expected = [ "foo" "bar" ];
+      };
     };
   cabalExecutableTests =
     let