+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - #215: Improved debug logging.
 - #216: Remove `debug` option (pass `--trace-verbose` to nix instead)
 - Breaking changes
-  - #221: Switch to `buildFromSdist`, to allow using non-standard package sets (wherein `cabal-install` is built without using user's overrides)
+  - #221: Switch to `buildFromSdist`, to allow using non-standard package sets (wherein `cabal-install` is otherwise built without using user's overrides)
 
 ## 0.4.0 (Aug 22, 2023)