+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - #210: Add `extraLibraries` to `settings` module.
 - #215: Improved debug logging.
 - #216: Remove `debug` option (pass `--trace-verbose` to nix instead)
+- Breaking changes
+  - #221: Switch to `buildFromSdist`, to allow using non-standard package sets (wherein `cabal-install` is otherwise built without using user's overrides)
 
 ## 0.4.0 (Aug 22, 2023)