+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased (0.5.0)
 
 - Breaking changes
-  - #221: Switch to `buildFromSdist`, to allow using non-standard package sets (wherein `cabal-install` is otherwise built without using user's overrides)
+  - #221: Switch from `buildFromCabalSdist` to `buildFromSdist`, to allow using non-standard package sets (wherein `cabal-install` is otherwise built without using user's overrides)
     - #253: Enable controlling `buildFromSdist` through `settings.<name>.buildFromSdist`. (This was turned off by default originally, but was turned on by default in #286)
 - Enhancements
   - `settings` module:
@@ -14,8 +14,7 @@
     - #216: Remove `debug` option (pass `--trace-verbose` to nix instead)
 - Fixes
   - #222: Improve `cabal.project` parser by handling files not ending with newline
-  - #223 Make `devShell.tools` a `lazyAttrsOf` (lazy evaluation of values)
-  - #271: Change all `types.attrsOf` to `types.lazyAttrsOf`. If you use `lib.mkIf` for `attrsOf` values (not `submodule` options), use `lib.optionalAttrs` instead. This fixes #270 (`basePackages`) and improves performance.
+  - #271, #223: Change all `types.attrsOf` to `types.lazyAttrsOf`. If you use `lib.mkIf` for `attrsOf` values (not `submodule` options), use `lib.optionalAttrs` instead. This fixes #270 (`basePackages`) and improves performance.
 
 ## 0.4.0 (Aug 22, 2023)